Refactoring the code
Initial refactor commit. Move the extensions out of the main VC into their own files, separated by type.
