+++ SlickStocks/Network/NetworkCalls.swift
@@ -1,0 +1,14 @@
+//
+//  Network.swift
+//  SlickStocks
+//
+//  Created by Adam Hogan on 3/6/18.
+//  Copyright © 2018 Adam Hogan. All rights reserved.
+//
+
+import Foundation
+
+class NetworkCalls {
+    
+    
+}
