+++ SlickStocks/View Controllers/MainViewController.swift
@@ -148,43 +148,7 @@ class MainViewController: UIViewController, UITableViewDelegate, UITableViewData
     }
 }
 
-extension String {
-    func convertToDecimal() -> Decimal {
-        let formatter = NumberFormatter()
-        formatter.generatesDecimalNumbers = true
-        return formatter.number(from: self) as! Decimal
-    }
-}
 
-extension Decimal {
-    func roundDecimalCurrency() -> String {
-        let formatter = NumberFormatter()
-        formatter.maximumFractionDigits = 2
-        formatter.minimumFractionDigits = 2
-        formatter.locale = Locale.current
-        formatter.numberStyle = .currency
-        return formatter.string(from: self as NSDecimalNumber)!
-    }
-}
 
-extension Decimal {
-    func roundDecimal() -> String {
-        let formatter = NumberFormatter()
-        formatter.maximumFractionDigits = 2
-        formatter.minimumFractionDigits = 2
-        formatter.numberStyle = .decimal
-        return formatter.string(from: self as NSDecimalNumber)!
-    }
-}
-
-extension Decimal {
-    func decimalNoFractions() -> String {
-        let formatter = NumberFormatter()
-        formatter.maximumFractionDigits = 0
-        formatter.minimumFractionDigits = 0
-        formatter.numberStyle = .decimal
-        return formatter.string(from: self as NSDecimalNumber)!
-    }
-}
